cleaned old files, added star rating system, fixed refetch queries for dishRatings, and connected firebase users with mongodb
