--- components/SignInWithGoogle.js
@@ -1,25 +1,0 @@
-import React from 'react'
-import { SafeAreaView, TouchableOpacity, Button } from 'react-native'
-import firebase from 'firebase/app'
-
-const SignInWithGoogle = ({ auth }) => {
-	const signInWithGoogle = () => {
-		const provider = new firebase.auth.GoogleAuthProvider()
-		auth.signInWithPopup(provider)
-	}
-
-	return (
-		<>
-			<SafeAreaView>
-				<TouchableOpacity>
-					<Button
-						onPress={signInWithGoogle}
-						title='Sign in with Google'
-					/>
-				</TouchableOpacity>
-			</SafeAreaView>
-
-		</>
-	)
-}
-export default SignInWithGoogle
