+++ package.json
@@ -35,7 +35,8 @@
     "react-native-reanimated": "~1.13.0",
     "react-native-safe-area-context": "3.1.9",
     "react-native-screens": "~2.15.2",
-    "react-native-svg": "^12.1.0",
+    "react-native-star-rating": "^1.1.0",
+    "react-native-svg": "^12.1.1-0",
     "react-native-vector-icons": "^8.0.0",
     "react-native-web": "~0.13.12",
     "react-navigation-bottom-tabs-no-warnings": "^5.11.3"
