+++ pages/Dish.js
@@ -1,8 +1,5 @@
 import React, { useState, useMemo, useEffect } from 'react';
-import { useMutation, useQuery } from '@apollo/client';
-import firebase from 'firebase/app';
-import { useAuthState } from 'react-firebase-hooks/auth';
-import { DISH_ADD_RATING, GET_DISH_RATINGS } from '../queries/dishes';
+import { useQuery } from '@apollo/client';
 import {
 	View,
 	StyleSheet,
@@ -10,33 +7,21 @@ import {
 	Image,
 	TouchableWithoutFeedback,
 	ScrollView,
-	Keyboard,
 	Text,
 } from 'react-native';
 import { sushi, back_arrow } from '../assets';
 import { Col, Row, Grid } from "react-native-easy-grid"
-import { 	Button, Card, Modal } from '@ui-kitten/components';
-import DishReviewRow from '../components/DishReviewRow';
+import { Button } from '@ui-kitten/components';
+
+import { GET_DISH_RATINGS } from '../queries/dishes';
+import DishReviewRow from '../components/Dish/DishReviewRow';
+import AddDishRatingModal from '../components/Dish/AddDishRatingModal';
 
 function Dish({ route }) {
-	const { navigation } = route.params
-	const { dish } = route.params
-	const userID = firebase.auth().currentUser.uid
+	const { navigation, dish } = route.params
 
-	const [rating, setRating] = useState('')
-	const [review, setReview] = useState('')
 	const [dishRatings, setDishRatings] = useState(0)
-	const [visible, setVisible] = React.useState(false);
-
-	// Adding a dishRating, then refetching all dishRatings to update the cache
-	const [addReview] = useMutation(DISH_ADD_RATING, {
-		onCompleted: (data) => {
-			console.log("Review successfully submitted", data);
-			Keyboard.dismiss()
-			refetch()
-		},
-		refetchQueries: [{ query: GET_DISH_RATINGS }]
-	})
+	const [isModalVisible, setModalVisible] = useState(false);
 
 	// Queries all dishRatings
 	const { loading, error, data, refetch } = useQuery(GET_DISH_RATINGS, {
@@ -44,9 +29,9 @@ function Dish({ route }) {
 			filter: { dish_id: dish._id, hasReviewText: true },
 			sort: '_ID_DESC'
 		},
-		onCompleted: (data) => {
-			console.log("Query Has Been Rerun:", data)
-		}
+		// onCompleted: (data) => {
+		// 	console.log("Query Has Been Rerun:", data)
+		// }
 	})
 
 	// Memoizes dishRatings and is updated when the dishRating Query is reran
@@ -54,20 +39,8 @@ function Dish({ route }) {
 		if (data) setDishRatings(data.dishRatingMany)
 	}, [data])
 
-	//Handler for adding dishRating
-	const addDishRatingHandler = () => {
-		addReview({
-			variables: {
-				record: {
-					dish_id: dish._id,
-					user_id: userID,
-					rating: Number(rating),
-					review,
-				}
-			}
-		})
-	}
-
+	if (loading) return <Text>Loading...</Text>
+	if (error) return <Text>Ratings had trouble loading, whoopsy...</Text>
 	return (
 		<View style={styles.container}>
 			<View style={styles.item_container}>
@@ -101,47 +74,22 @@ function Dish({ route }) {
 								<Text style={styles.review_title}>Reviews</Text>
 							</Col>
 							<Col>
-								<Button style={styles.add_review_btn} onPress={() => setVisible(true)}>
+								<Button style={styles.add_review_btn} onPress={() => setModalVisible(true)}>
 									Add Review
-      </Button>
+      							</Button>
 							</Col>
 						</Row>
-						<DishReviewRow dish = {dish}/>
+						<DishReviewRow dish={dish} />
 					</Grid>
 				</ScrollView>
 			</View>
+			<AddDishRatingModal
+				isModalVisible={isModalVisible}
+				setModalVisible={setModalVisible}
+				dish_id={dish._id}
+				refetchAllRatings={refetch}
+			/>
 
-			<Modal
-				visible={visible}
-				backdropStyle={styles.backdrop}
-				onBackdropPress={() => setVisible(false)}>
-				<Card disabled={true}>
-					<Row>
-						<TextInput
-							style={{ height: 50, borderColor: 'gray', borderWidth: 1 }}
-							onChangeText={text => setReview(text)}
-							value={review}
-							placeholder='Add Review Text Here'
-							style={{}}
-						/>
-					</Row>
-					<Row>
-						<TextInput
-							style={{ height: 40, borderColor: 'gray', borderWidth: 1 }}
-							onChangeText={num => setRating(num)}
-							value={rating}
-							keyboardType='numeric'
-							placeholder='Add Rating Number Here'
-							style={{ marginTop: 30 }}
-						/>
-					</Row>
-					<Row>
-					</Row>
-					<Button title="Add Review" onPress={addDishRatingHandler}>
-						Add Review
-          </Button>
-				</Card>
-			</Modal>
 		</View>
 	);
 }
@@ -150,9 +98,6 @@ const styles = StyleSheet.create({
 	container: {
 		flex: 1,
 		backgroundColor: '#FDFCFC',
-	},
-	backdrop: {
-		backgroundColor: 'rgba(0, 0, 0, 0.5)',
 	},
 	back_arrow: {
 		marginTop: '10%',
